requirements.txt
pycryptodome==3.10.1
python-dotenv==0.19.0
